homework4.2 Sort Bubble
